[front/components/agent_builder/observability/charts] - refactor: consolidate tooltip components for charts
- Extracted common tooltip logic into a single `ChartsTooltip` component to be used across different chart types
- Removed redundant `FeedbackDistributionTooltip` and `ToolUsageTooltip` components in favor of the unified `ChartsTooltip`
- Updated `FeedbackDistributionChart` and `ToolUsageChart` to utilize the new `ChartsTooltip` component for displaying tooltips
- Ensured tooltips display relevant information based on chart mode and type while maintaining existing functionality
